pyproject.toml
[tool.poetry]
name = "vidyaNav-ai-backend"
version = "0.1.0"
description = "Sahayak - AI teaching assistant for multi-grade students in rural area of India"
authors = ["AnishNavalgund <anishk.navalgund@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
google-cloud-aiplatform = ">=1.38.0,<2.0.0"
google-cloud-storage = ">=2.12.0,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
flask = ">=3.1.1,<4.0.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-multipart = "^0.0.20"
pydantic-ai = {version = "^0.4.2", extras = ["vertexai"]}
pydantic-ai-slim = {extras = ["google"], version = "^0.4.2"}
google-cloud-texttospeech = "^2.27.0"
langchain = "^0.3.26"
pypdf = "^5.8.0"
vertexai = "^1.71.1"
qdrant-client = "^1.14.3"
langchain-community = "^0.3.27"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"